Update is-land-a-big-deal.md
Ok, I finally finished trying to update the language in the first part of the article, I decided on contrasting "rental value" with "selling value". I'm not sure if adding brackets for land (selling) value breaks up the flow too much or if removing the brackets would be better. 

In general, this was a lot harder than I thought it'd be. "value" is used in different ways throughout the text "property value, real estate value, land value, market value, full market value, etc" that it was hard to decide on a strict "conversion" method. My hope is that my changes will reduce confusion instead of increasing it but I'm not completely sure.